agentic-safety-shell/pyproject.toml
[project]
name = "agentic-safety-shell"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.14"
dependencies = [
    "google-genai>=1.64.0",
    "google-generativeai>=0.8.6",
    "python-dotenv>=1.2.1",
]
